+++ R/PipeOpADAS.R
@@ -5,7 +5,7 @@
 #' @format [`R6Class`][R6::R6Class] object inheriting from [`PipeOpTaskPreproc`]/[`PipeOp`].
 #'
 #' @description
-#' Generates a more balanced data set by creating synthetic instances of the minority class using the ADASYN algorithm.
+#' Generates a more balanced data set by creating synthetic instances of the minority classes using the ADASYN algorithm.
 #'
 #' The algorithm generates for each minority instance new data points based on its `K` nearest neighbors and the difficulty of learning for that data point.
 #' It can only be applied to tasks with numeric features that have no missing values.
@@ -37,6 +37,10 @@
 #' * `K` :: `numeric(1)` \cr
 #'   The number of nearest neighbors used for sampling new values. Default is `5`.
 #'   See [`ADAS()`][`smotefamily::ADAS`].
+#'
+#' @section Internals:
+#' If a target level is unobserved during training, no synthetic data points will be generated for that class.
+#' No error is raised; the unobserved class is simply ignored.
 #'
 #' @section Fields:
 #' Only fields inherited from [`PipeOp`].
@@ -98,7 +102,11 @@ PipeOpADAS = R6Class("PipeOpADAS",
 
       # Calculate synthetic data
       dt = task$data(cols = cols)
-      st = setDT(invoke(smotefamily::ADAS, X = dt, target = task$truth(),
+      # Remove unseen target levels, see #881
+      # Don't need to re-add them later since we don't touch task here
+      target = droplevels(task$truth())
+
+      st = setDT(invoke(smotefamily::ADAS, X = dt, target = target,
         .args = self$param_set$get_values(tags = "adas"),
         .opts = list(warnPartialMatchArgs = FALSE))$syn_data)  # ADAS uses partial arg matching internally